requirements.txt
pandas
numpy
scikit-learn
matplotlib
pyyaml
scipy
jupyter